[level 0] Title: 구슬을 나누는 경우의 수, Time: 1.02 ms, Memory: 75.4 MB -BaekjoonHub

+++ Java/프로그래머스/0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.java
@@ -1,6 +1,6 @@
 import java.math.BigInteger;
 class Solution {
-  public long solution(int balls, int share) {
+  public int solution(int balls, int share) {
     if(balls == share) return 1;
     if(share == 1) return balls;
     if(balls - share == 1) return balls;
@@ -10,6 +10,6 @@ class Solution {
       arr[i] = new BigInteger(String.valueOf(i)).multiply(arr[i - 1]);
     }
     BigInteger answer = arr[balls].divide(arr[balls - share]).divide(arr[share]);
-    return answer.longValue();
+    return answer.intValue();
   }
 }
